+++ Source/Core/Geometry/Geometry.hpp
@@ -92,7 +92,7 @@ void ComputeBestFitPlane(const Vec3* polyPoints, uint polyPointCount, Vec3Ptr pl
 /// assumes a scale of (1, 1, 1).
 real CalculateTriMeshVolume(const Vec3* triMeshPoints,
                             const uint* triMeshTriangles,
-                            uint triangleCount,
+                            size_t triangleCount,
                             Vec3Param scale = NoScale);
 real CalculateTriMeshVolume(const Array<Vec3>& triMeshPoints,
                             const Array<uint>& triMeshTriangles,
