+++ Source/Physics/ConstraintAtoms.hpp
@@ -15,8 +15,8 @@ struct AnchorAtom
 {
   AnchorAtom();
 
-  Vec3Ref operator[](uint index);
-  Vec3Param operator[](uint index) const;
+  Vec3Ref operator[](size_t index);
+  Vec3Param operator[](size_t index) const;
 
   Vec3 mBodyR[2];
 };
@@ -31,11 +31,11 @@ struct WorldAnchorAtom
 
   void SetUp(const AnchorAtom& anchor, Collider* obj1, Collider* obj2);
 
-  inline Vec3& operator[](uint index)
+  inline Vec3& operator[](size_t index)
   {
     return mWorldR[index];
   };
-  inline const Vec3& operator[](uint index) const
+  inline const Vec3& operator[](size_t index) const
   {
     return mWorldR[index];
   };
@@ -54,11 +54,11 @@ struct AxisAtom
   {
     mBodyAxes[0] = mBodyAxes[1] = Vec3::cYAxis;
   };
-  inline Vec3& operator[](uint index)
+  inline Vec3& operator[](size_t index)
   {
     return mBodyAxes[index];
   };
-  inline const Vec3& operator[](uint index) const
+  inline const Vec3& operator[](size_t index) const
   {
     return mBodyAxes[index];
   };
@@ -73,11 +73,11 @@ struct WorldAxisAtom
   WorldAxisAtom(){};
   WorldAxisAtom(const AxisAtom& axes, Collider* obj1, Collider* obj2);
 
-  inline Vec3& operator[](uint index)
+  inline Vec3& operator[](size_t index)
   {
     return mWorldAxes[index];
   };
-  inline const Vec3& operator[](uint index) const
+  inline const Vec3& operator[](size_t index) const
   {
     return mWorldAxes[index];
   };
@@ -95,11 +95,11 @@ struct AngleAtom
     mLocalAngles[1] = Quat::cIdentity;
   };
 
-  inline Quat& operator[](uint index)
+  inline Quat& operator[](size_t index)
   {
     return mLocalAngles[index];
   };
-  inline const Quat& operator[](uint index) const
+  inline const Quat& operator[](size_t index) const
   {
     return mLocalAngles[index];
   };
@@ -116,11 +116,11 @@ struct WorldAngleAtom
   WorldAngleAtom();
   WorldAngleAtom(const AngleAtom& refAngle, Collider* obj1, Collider* obj2);
 
-  inline Quat& operator[](uint index)
+  inline Quat& operator[](size_t index)
   {
     return mWorldAngles[index];
   };
-  inline const Quat& operator[](uint index) const
+  inline const Quat& operator[](size_t index) const
   {
     return mWorldAngles[index];
   };
